requirements.txt
streamlit>=1.24
mysql-connector-python>=8.0
pandas
